packages/python-packages/nixoslogo/pyproject.toml
[project]
name = "nixoslogo"
version = "0.1.0"
description = ""
authors = [
    {name = "Daniel Baker",email = "daniel.n.baker@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "svg-py (>=1.6.0,<2.0.0)",
    "coloraide (>=4.4.1,<5.0.0)",
    "jsonpickle (>=4.0.2,<5.0.0)"
]

[tool.poetry]
packages = [{include = "nixoslogo"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
